+++ 1.20/Common/src/main/java/fuzs/puzzleslib/api/core/v1/ModConstructor.java
@@ -13,7 +13,17 @@ import java.util.function.Supplier;
 /**
  * a base class for a mods main common class, contains a bunch of methods for registering various things
  */
-public interface ModConstructor extends PairedModConstructor {
+public interface ModConstructor extends BaseModConstructor {
+
+    /**
+     * Construct the main {@link ModConstructor} instance provided as <code>supplier</code> to begin initialization of a mod.
+     *
+     * @param modId    the mod id for registering events on Forge to the correct mod event bus
+     * @param supplier the main mod instance for mod setup
+     */
+    static void construct(String modId, Supplier<ModConstructor> supplier) {
+        construct(modId, supplier, new ContentRegistrationFlags[0]);
+    }
 
     /**
      * Construct the main {@link ModConstructor} instance provided as <code>supplier</code> to begin initialization of a mod.
@@ -21,18 +31,21 @@ public interface ModConstructor extends PairedModConstructor {
      * @param modId    the mod id for registering events on Forge to the correct mod event bus
      * @param supplier the main mod instance for mod setup
      * @param flags    specific content this mod uses that needs to be additionally registered
+     * @deprecated replaced provided flags with {@link BaseModConstructor#getContentRegistrationFlags()}
      */
+    @Deprecated(forRemoval = true)
     static void construct(String modId, Supplier<ModConstructor> supplier, ContentRegistrationFlags... flags) {
-        if (Strings.isBlank(modId)) throw new IllegalArgumentException("mod id must not be empty");
+        if (Strings.isBlank(modId)) throw new IllegalArgumentException("mod id is empty");
         // build first to force class being loaded for executing buildables
-        ModConstructor modConstructor = supplier.get();
-        ResourceLocation identifier = ModContext.getPairingIdentifier(modId, modConstructor);
+        ModConstructor instance = supplier.get();
+        ResourceLocation identifier = ModContext.getPairingIdentifier(modId, instance);
         PuzzlesLib.LOGGER.info("Constructing common components for {}", identifier);
         ModContext modContext = ModContext.get(modId);
-        Set<ContentRegistrationFlags> availableFlags = Set.of(flags);
+        ContentRegistrationFlags[] builtInFlags = instance.getContentRegistrationFlags();
+        Set<ContentRegistrationFlags> availableFlags = Set.of(builtInFlags.length != 0 ? builtInFlags : flags);
         Set<ContentRegistrationFlags> flagsToHandle = modContext.getFlagsToHandle(availableFlags);
         modContext.beforeModConstruction();
-        CommonFactories.INSTANCE.constructMod(modId, modConstructor, availableFlags, flagsToHandle);
+        CommonFactories.INSTANCE.constructMod(modId, instance, availableFlags, flagsToHandle);
         modContext.afterModConstruction(identifier);
     }
 
@@ -48,7 +61,6 @@ public interface ModConstructor extends PairedModConstructor {
      * used to set various values and settings for already registered content
      *
      * @param context enqueue work to be run sequentially for all mods as the setup phase runs in parallel on Forge
-     *
      * @deprecated now always runs deferred, use {@link #onCommonSetup()}
      */
     default void onCommonSetup(final ModLifecycleContext context) {