+++ 1.20/Fabric/src/main/java/fuzs/puzzleslib/impl/core/FabricModConstructor.java
@@ -2,9 +2,9 @@ package fuzs.puzzleslib.impl.core;
 
 import fuzs.puzzleslib.api.core.v1.ContentRegistrationFlags;
 import fuzs.puzzleslib.api.core.v1.ModConstructor;
-import fuzs.puzzleslib.impl.item.CopyTagRecipe;
 import fuzs.puzzleslib.api.item.v2.LegacySmithingTransformRecipe;
 import fuzs.puzzleslib.impl.core.context.*;
+import fuzs.puzzleslib.impl.item.CopyTagRecipe;
 import net.minecraft.core.Registry;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.ResourceLocation;
@@ -20,7 +20,7 @@ public final class FabricModConstructor {
 
     public static void construct(ModConstructor constructor, String modId, Set<ContentRegistrationFlags> availableFlags, Set<ContentRegistrationFlags> flagsToHandle) {
         registerContent(modId, flagsToHandle);
-        registerHandlers(constructor, modId);
+        registerHandlers(constructor, modId, availableFlags);
     }
 
     private static void registerContent(String modId, Set<ContentRegistrationFlags> flagsToHandle) {
@@ -34,7 +34,7 @@ public final class FabricModConstructor {
         }
     }
 
-    private static void registerHandlers(ModConstructor constructor, String modId) {
+    private static void registerHandlers(ModConstructor constructor, String modId, Set<ContentRegistrationFlags> availableFlags) {
         constructor.onConstructMod();
         constructor.onRegisterCreativeModeTabs(new CreativeModeTabContextFabricImpl());
         constructor.onBuildCreativeModeTabContents(new BuildCreativeModeTabContentsContextFabricImpl());
@@ -44,7 +44,7 @@ public final class FabricModConstructor {
         constructor.onRegisterSpawnPlacements(new SpawnPlacementsContextFabricImpl());
         constructor.onRegisterFuelBurnTimes(new FuelBurnTimesContextFabricImpl());
         constructor.onRegisterFlammableBlocks(new FlammableBlocksContextFabricImpl());
-        constructor.onRegisterBiomeModifications(new BiomeModificationsContextFabricImpl(modId));
+        constructor.onRegisterBiomeModifications(new BiomeModificationsContextFabricImpl(modId, availableFlags));
         constructor.onAddDataPackFinders(new DataPackSourcesContextFabricImpl());
         constructor.onRegisterDataPackReloadListeners(new AddReloadListenersContextFabricImpl(PackType.SERVER_DATA, modId));
     }